fix(content-sidebar): Fixed circular dependency for box ai sidebar
Apply suggestions from code review
